Move math routines in _math module
